[perf] 렌더링 최적화 - 룰렛 애니메이션 렌더링 최적화 (#711)
* refactor : 룰렛 휠 렌더링 문제 해결

* refactor : AnimatedRouletteWheel 컴포넌트로 따로 분리
