+++ frontend/src/features/roulette/components/RouletteWheel/RouletteWheel.styled.ts
@@ -7,10 +7,6 @@ type WrapperProps = {
 };
 
 export const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
   position: relative;
 `;
